+++ backend/bootstrap/urls.py
@@ -28,9 +28,10 @@ router = routers.SimpleRouter(trailing_slash=False)
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('api/', include(("api.urls", "api"), namespace="api")),
+    # path('api/', include(("api.urls", "api"), namespace="api")),
     path('admin/', admin.site.urls),
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
-    path('dj-rest-auth/', include('dj_rest_auth.urls')),
-    path('accounts/', include('allauth.urls')),
+    # path('dj-rest-auth/', include('dj_rest_auth.urls')),
+    # path('accounts/', include('allauth.urls')),
+    path('auth/', include('allauth.headless.urls')),
 ]
