+++ backend/core/utils/common.py
