+++ backend/core/utils/fixtures.py
