+++ backend/env.sample
@@ -3,7 +3,7 @@ ALLOWED_HOSTS=localhost,127.0.0.1,[::1]
 SECRET_KEY=django-insecure-728k0bs%91o$^sp%aa_ji@2fmtwpdk7r1na#*$%l2+%)7tnpo3
 
 ##########################################################
-# DATABASE
+#  DATABASE
 ##########################################################
 
 DB_ENGINE=django.db.backends.postgresql
@@ -17,11 +17,16 @@ DB_PORT=5432
 #  CORS
 ##########################################################
 
-# CORS_ALLOW_ALL_ORIGINS=True
-# CORS_ALLOWED_ORIGINS=
+# CORS_ALLOW_ALL_ORIGINS=False
+# CORS_ALLOWED_ORIGINS=http://localhost:3000
+# CORS_ALLOW_CREDENTIALS=False
+
+#  CSRF
+# CSRF_TRUSTED_ORIGINS=http://localhost:3000
+
 
 #########################################################
-# ALLAUTH
+#  ALLAUTH
 #########################################################
 
 GOOGLE_CLIENT_ID=
