+++ backend/search/viewsets.py
@@ -4,7 +4,7 @@ from rest_framework import generics, viewsets, status
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from utils.io import IOMixin, paginate_response
+from core.utils.io import IOMixin, paginate_response
 from core.models import Issue
 from core.serializers import IssueSerializer
 from .search import IssueSearch
